✨ 새로운 컬럼 추가하기 완. 그런데 자동으로 추가된 컬럼 렌더링이 안됨
