+++ src/components/add-column-button/index.tsx
@@ -1,14 +1,21 @@
 import { ReactNode } from 'react';
 import Button from './style';
+import { useDispatch } from 'react-redux';
+import { openModal } from '@/redux/modalSlice';
 
 interface Props {
   children: ReactNode;
-  createColumns: () => void;
+  createColumns?: () => void;
 }
 
-const AddColumnButton = ({ children, createColumns }: Props) => {
+const AddColumnButton = ({ children }: Props) => {
+  const dispatch = useDispatch();
+  const handleButtonClick = () => {
+    dispatch(openModal('addColumnModal'));
+  };
+
   return (
-    <Button type='button' onClick={createColumns}>
+    <Button type='button' onClick={handleButtonClick}>
       {children}
       <img src='/assets/image/icons/bannerAddIcon.svg' alt='add-icon' />
     </Button>
